Add files via upload
The goal of the project was to use Ansible to automate setup and configuration to increase the management of a large-scale infrastructure's effectiveness and consistency. The project's objectives were to build an automated and scalable infrastructure provided by automatically provisioning servers, installing and configuring software, and managing and maintaining the infrastructure, all with minimal manual intervention.

The project's desired result was a completely automated infrastructure that could be scaled and controlled using Ansible playbooks with little to no manual labor. By using these playbooks to provision servers, install and configure software, and carry out ongoing maintenance operations, the infrastructure can be managed more effectively and consistently, requiring less manual involvement.